Traitement du formulaire

+++ android_client/src/main/java/classes/model/Language.java
@@ -15,11 +15,11 @@ public class Language {
     @Element
     private String name;
     @Element
-    private int level;
+    private float level;
 
     public final static int MAX_LEVEL = 20;
 
-    public Language(String name, int level) {
+    public Language(String name, float level) {
         this.name = name;
         setLevel(level);
     }
@@ -37,11 +37,11 @@ public class Language {
         this.name = name;
     }
 
-    public int getLevel() {
+    public float getLevel() {
         return level;
     }
 
-    public void setLevel(int level) {
+    public void setLevel(float level) {
         if (level < 0) {
             level = 0;
         } else if (level > MAX_LEVEL) {
