+++ android_client/src/main/java/classes/model/Skill.java
@@ -14,7 +14,7 @@ public class Skill {
     @Element
     private String name;
     @Element
-    private int level;
+    private float level;
     public static final int MAX_LEVEL = 20;
 
     public String getName() {
@@ -25,7 +25,7 @@ public class Skill {
         this.name = name;
     }
 
-    public Skill(String name, int level) {
+    public Skill(String name, float level) {
         this.name = name;
         setLevel(level);
     }
@@ -34,11 +34,11 @@ public class Skill {
 
     }
 
-    public int getLevel() {
+    public float getLevel() {
         return level;
     }
 
-    public void setLevel(int level) {
+    public void setLevel(float level) {
         if (level < 0) {
             this.level = 0;
         } else if (level > MAX_LEVEL) {
